move root tests, #420
