+++ tests/synthetic_data.py
