+++ tests/test_metastandards.py
