+++ tests/test_trends_engine.py
