+++ tests/test_variable.py
